Room 입장 수정

+++ sharedDatabase/src/commonMain/kotlin/com/wonddak/database/queriesHelper/RoomInfoQueriesHelper.kt
@@ -29,6 +29,8 @@ class RoomInfoQueriesHelper(
     }
 
     fun deleteRoomInfo(roomId: String) {
-        queries.deleteById(roomId)
+        if (roomId.isNotEmpty()) {
+            queries.deleteById(roomId)
+        }
     }
 }
